pyproject.toml
[project]
name = "streamlit-profile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.1.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "openai>=2.26.0",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
